Draft 2 with admin controller and modal
