+++ src/controllers/logoutAdminController.js
@@ -1,0 +1,16 @@
+import { useNavigate } from 'react-router-dom';
+import { logoutUser } from '../models/logoutService.js';
+
+const useAdminLogoutController = () => {
+  const navigate = useNavigate();
+
+  const handleLogout = async () => {
+    await logoutUser();
+    navigate('/admin-login');
+    
+  };
+
+  return {handleLogout};
+};
+
+export default useAdminLogoutController;
